+++ scalr/data_source_environment.go
@@ -104,8 +104,8 @@ func dataSourceEnvironmentRead(d *schema.ResourceData, meta interface{}) error {
 	}
 
 	if err != nil {
-		if errors.Is(err, scalr.ErrResourceNotFound{}) {
-			return fmt.Errorf("Environment %s not found", envID)
+		if errors.Is(err, scalr.ErrResourceNotFound) {
+			return fmt.Errorf("Environment '%s' not found", envID)
 		}
 		return fmt.Errorf("Error retrieving environment: %v", err)
 	}