+++ scalr/resource_scalr_webhook.go
@@ -192,7 +192,7 @@ func resourceScalrWebhookRead(d *schema.ResourceData, meta interface{}) error {
 	log.Printf("[DEBUG] Read endpoint with ID: %s", webhookID)
 	webhook, err := scalrClient.Webhooks.Read(ctx, webhookID)
 	if err != nil {
-		if errors.Is(err, scalr.ErrResourceNotFound{}) {
+		if errors.Is(err, scalr.ErrResourceNotFound) {
 			return fmt.Errorf("Could not find webhook %s: %v", webhookID, err)
 		}
 		return fmt.Errorf("Error retrieving webhook: %v", err)
@@ -255,7 +255,7 @@ func resourceScalrWebhookDelete(d *schema.ResourceData, meta interface{}) error 
 	log.Printf("[DEBUG] Delete webhook: %s", d.Id())
 	err := scalrClient.Webhooks.Delete(ctx, d.Id())
 	if err != nil {
-		if errors.Is(err, scalr.ErrResourceNotFound{}) {
+		if errors.Is(err, scalr.ErrResourceNotFound) {
 			return nil
 		}
 		return fmt.Errorf("Error deleting webhook %s: %v", d.Id(), err)