+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed ([[#117]](https://github.com/Scalr/terraform-provider-scalr/pull/117))
+
+- Correctly handle not found resources
+
+
 ## [1.0.0-rc27] - 2022-02-17
 
 ### Fixed